requirements.txt
asgiref==3.6.0
cloudinary==1.33.0
dj-database-url==0.5.0
dj3-cloudinary-storage==0.0.6
Django==3.2.19
django-allauth==0.54.0
django-appconf==1.0.5
django-braces==1.15.0
django-cloudinary-storage==0.3.0
django-crispy-forms==1.14.0
django-etc==1.4.0
django-hitcount==1.3.5
django-imagekit==4.1.0
django-js-asset==2.0.0
django-model-utils==4.3.1
django-resized==1.0.2
django-summernote==0.8.20.0
django-taggit==4.0.0
djangorestframework==3.14.0
gunicorn==20.1.0
oauthlib==3.2.2
pilkit==2.0
Pillow==9.5.0
psycopg2==2.9.6
PyJWT==2.7.0
python3-openid==3.2.0
pytz==2023.3
requests-oauthlib==1.3.1
sqlparse==0.4.4
swapper==1.3.0
urllib3==1.26.15
